+++ samples/elaborate/meta_sample.rb
@@ -264,7 +264,7 @@ class MetaSampleApplication
           }
         }
             
-        @code_text = code_text {
+        @code_text = code_text(lines: {width: 2}) {
           text bind(SampleDirectory, 'selected_sample.code', read_only: true)
           editable bind(SampleDirectory, 'selected_sample.editable')
         }